TG-3: improve modal with new functionalities for delete
