+++ src/Components/Tasks/Modal/modal.module.css
@@ -1,15 +1,36 @@
 .container {
-    display: flex;
+    position: fixed;
+    left: 0;
+    top: 0;
+    right: 0;
+    bottom: 0;
     justify-content: center;
     align-items: center;
-    width: 100%;
-    height: 100%;
     background-color: rgba(0,0,0,0.5);
 }
 
 .modal {
-    width: 300px;
-    height: 100px;
+    width: 500px;
+    height: 150px;
     background-color: white;
     border-radius: 5px;
+    border: solid 1px grey;
+    margin: auto;
+    margin-top: 200px;
+}
+
+.buttonConfirm {
+    padding: 10px 24px;
+    background: #F13312;
+    border-radius: 5px;
+    color: white;
+    margin-left: 10px;
+}
+
+.buttonClose {
+    padding: 10px 24px;
+    background: #AACE9B;
+    border-radius: 5px;
+    color: white;
+    margin-left: 10px;
 }
