+++ src/Components/Tasks/Table/table.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Row from '../Row/row';
 import styles from './table.module.css';
 
-const Table = ({ tasks, openModal }) => {
+const Table = ({ tasks, handleDelete }) => {
   return (
     <table className={styles.table}>
       <thead>
@@ -13,11 +13,20 @@ const Table = ({ tasks, openModal }) => {
         </tr>
       </thead>
       <tbody>
-        {tasks.map((task) => {
-          return <Row key={task._id} task={task} openModal={openModal} />;
-        })}
+        {tasks === undefined || tasks.length === 0 ? (
+          <td>There are no tasks yet!</td>
+        ) : (
+          tasks.map((task) => {
+            return <Row key={task._id} task={task} handleDelete={handleDelete} />;
+          })
+        )}
       </tbody>
     </table>
+    // {tasks === undefined || tasks.length === 0 ? (
+    //   <h3 className={styles.container}>There are no tasks yet!</h3>
+    // ) : (
+    //   <Table tasks={tasks} handleDelete={handleDelete} />
+    // )}
   );
 };
