+++ src/Components/Tasks/Table/table.module.css
@@ -14,4 +14,5 @@
   .table td,th {
     text-align: center;
     padding: 12px 15px;
+    color: #373867;
   }
