+++ src/Components/Tasks/index.js
@@ -2,10 +2,8 @@ import React from 'react';
 import { useEffect, useState } from 'react';
 import Table from './Table/table';
 import styles from './tasks.module.css';
-import Modal from './Modal/modal-delete';
 
 function Tasks() {
-  const [showModal, setShowModal] = useState(false);
   const [tasks, saveTasks] = useState([]);
 
   useEffect(() => {
@@ -21,31 +19,16 @@ function Tasks() {
       method: 'DELETE'
     });
     saveTasks([...tasks.filter((newListItem) => newListItem._id !== id)]);
-    closeModal();
   };
 
-  if (tasks === undefined) {
-    return <p className={styles.container}>There are no tasks yet!</p>;
-  }
-
-  const closeModal = () => {
-    setShowModal(false);
-  };
-
-  const openModal = () => {
-    setShowModal(true);
-  };
+  // if (tasks === undefined) {
+  //   return <h3 className={styles.container}>There are no tasks yet!</h3>;
+  // }
 
   return (
-    <div>
-      <Modal
-        showModal={showModal}
-        openModal={openModal}
-        closeModal={closeModal}
-        handleDelete={handleDelete}
-        tasks={tasks}
-      />
-      <Table tasks={tasks} openModal={openModal} />
+    <div className={styles.container}>
+      <h2>Tasks</h2>
+      <Table tasks={tasks} handleDelete={handleDelete} />
     </div>
   );
 }
