+++ src/Components/Tasks/tasks.module.css
@@ -3,4 +3,5 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  flex-direction: column;
 }
